restructure flamethrower file, add flamethrower damage
