+++ Assets/Scripts/FlamethrowerBehavior.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8137669aee428344193d207425d4f24b
+guid: d1c257a9567f8a249a966dbc2fcb9304
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
